requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
httpx==0.27.2
PyJWT==2.9.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1

